Module Mm.Image

Operations on images.

exception Invalid_position

Trying to access a pixel outside the boundaries of the image.

exception Invalid_dimensions

The image does not have the expected dimensions.

module Data : sig ... end

module Pixel : sig ... end

module Point : sig ... end

module Fraction : sig ... end

module Draw : sig ... end

module Bitmap : sig ... end

Operation on bitmaps, which are black and white images.

module RGB8 : sig ... end

Operations on images stored in RGB8 format, ie RGB channels, one byte each.

module ARGB8 : sig ... end

module BGRA : sig ... end

module RGBA32 : sig ... end

Operations on images stored in RGBA32 format (ie RGB channels + an alpha channel, one byte for each).

module YUV420 : sig ... end

Operations on images stored in YUV420 format, ie one luma (Y) and two chrominance (U and V) channels.

module Generic : sig ... end

Operations on images in generic formats (many formats are supported).

module type CanvasImage = sig ... end

Type of image module expected to build canvas.

module Canvas (I : CanvasImage) : sig ... end

Canvas of images, i.e. formal sums of images of various dimensions with various offsets.

module CanvasYUV420 : sig ... end