Class type Reader.t

method read : int -> Multitrack.buffer -> int -> int -> int

Read data at with given samplerate for events, in a given track, with a given length in samples.

method close : unit

Close the stream.