Class Synth.sine

Sine synthesizer.

method set_volume : float -> unit

Set the global volume of the synth.

method note_on : int -> float -> unit

Play a note.

method note_off : int -> float -> unit

Stop playing a note.

method fill_add : Mm_audio.Audio.buffer -> int -> int -> unit

Fill a buffer with synthesized data adding to the original data of the buffer.

method play : Mm_midi.MIDI.buffer -> int -> Mm_audio.Audio.buffer -> int -> int -> unit

Synthesize into an audio buffer. Notice that the delta times in the track should be in samples (so they do depend on the samplerate).

method play_add : Mm_midi.MIDI.buffer -> int -> Mm_audio.Audio.buffer -> int -> int -> unit

Same as play but keeps data originally present in the buffer.

method reset : unit

Reset the synthesizer (sets all notes off in particular).