Module Audio.Buffer_ext

Buffers of variable size. These are particularly useful for temporary buffers.

type t

val create : int -> int -> t

Create an extensible buffer of given channels and initial size in samples.

val prepare : t -> ?channels:int -> int -> buffer

Make sure that the buffer can hold at least a given number of samples.