Module Audio.Mono

Operations on mono buffers (with only one channel).

type t = float array

A mono buffer.

type buffer = t

val create : int -> t

val make : int -> float -> t

val sub : t -> int -> int -> t

val blit : t -> int -> t -> int -> int -> unit

val copy : t -> int -> int -> t

val copy_to_ba : t -> int -> int -> (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> unit

val copy_from_ba : (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> t -> int -> int -> unit

val of_ba : (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> t

val to_ba : t -> int -> int -> (float, Stdlib.Bigarray.float32_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val copy_to_int16_ba : t -> int -> int -> (int, Stdlib.Bigarray.int16_signed_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> unit

val copy_from_int16_ba : (int, Stdlib.Bigarray.int16_signed_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> t -> int -> int -> unit

val of_int16_ba : (int, Stdlib.Bigarray.int16_signed_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t -> t

val to_int16_ba : t -> int -> int -> (int, Stdlib.Bigarray.int16_signed_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val length : t -> int

Length in samples.

val append : t -> int -> int -> t -> int -> int -> t

val clear : t -> int -> int -> unit

Clear a portion of a buffer (fill it with zeroes).

val amplify : float -> t -> int -> int -> unit

val resample : ?mode:[ `Nearest | `Linear ] -> float -> t -> int -> int -> t

val clip : t -> int -> int -> unit

val noise : t -> int -> int -> unit

val squares : t -> int -> int -> float

val add : t -> int -> t -> int -> int -> unit

Samplewise add two buffers, storing the result in the first one.

val mult : t -> int -> t -> int -> int -> unit

Samplewise multiply two buffers of the same length, storing the result in the first one.

module Ringbuffer_ext : Mm_base.Ringbuffer.R with type buffer = t

module Ringbuffer : Mm_base.Ringbuffer.R with type buffer = t

module Buffer_ext : sig ... end

Buffers of variable size. These are particularly useful for temporary buffers.

module Analyze : sig ... end

Functions for analyzing audio data.

module Effect : sig ... end

module Generator : sig ... end

Sound generators.