Module FFT.Window

Windowing functions. These can be used to on complex buffers in order to improve the quality of the FFT, see http://en.wikipedia.org/wiki/Windowing_functions.

val cosine : Stdlib.Complex.t array -> unit

val hann : Stdlib.Complex.t array -> unit

val hamming : Stdlib.Complex.t array -> unit

val lanczos : Stdlib.Complex.t array -> unit

val triangular : Stdlib.Complex.t array -> unit

val bartlett_hann : Stdlib.Complex.t array -> unit

val blackman : ?alpha:float -> Stdlib.Complex.t array -> unit

val nuttall : Stdlib.Complex.t array -> unit

val blackman_harris : Stdlib.Complex.t array -> unit

val blackman_nuttall : Stdlib.Complex.t array -> unit