Class type Generator.t

A sound generator.

method set_volume : float -> unit

method set_frequency : float -> unit

method fill : buffer -> unit

Fill a buffer with generated sound.

method fill_add : buffer -> unit

Same as fill but adds the sound to the buffer.

method release : unit

Release the generator (used for generator with envelopes).

method dead : bool

Is the generator still producing sound? This should become false soon after release has been triggered.