Module ImageBase.Data

type t = (int, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val alloc : int -> (int, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val round : int -> int -> int

round n k rounds n to the nearest upper multiple of k.

val aligned : int -> int -> t

aligned k n allocates n bytes at a multiple of k.

val rounded_plane : int -> int -> int * t

val to_string : t -> string

val to_bytes : t -> bytes

val of_string : string -> t

val blit_all : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> unit

val blit : t -> int -> t -> int -> int -> unit

val copy : t -> t

val sub : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> int -> int -> ('a, 'b, 'c) Stdlib.Bigarray.Array1.t

val length : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> int

val size : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> int

val get : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> int -> 'a

val fill : ('a, 'b, 'c) Stdlib.Bigarray.Array1.t -> 'a -> unit