Module Mm_image.ImageCanvas

module type CanvasImage = sig ... end

module Canvas (I : CanvasImage) : sig ... end

A canvas of images. The structure is immutable but its elements might be returned and therefore should not be used in place.

module CanvasYUV420 : sig ... end