Module ImageCanvas.Canvas

A canvas of images. The structure is immutable but its elements might be returned and therefore should not be used in place.

Parameters

module I : CanvasImage

Signature

module Element : sig ... end

module E = Element

type t = {

width : int;

height : int;

elements : E.t list;

}

val create : int -> int -> t

val width : t -> int

val height : t -> int

val size : t -> int

val make : ?width:int -> ?height:int -> ?x:int -> ?y:int -> I.t -> t

val add : t -> t -> t

val covering : t -> bool

val render : ?fresh:bool -> ?transparent:bool -> t -> I.t

val rendered : ?transparent:bool -> t -> t

val map : ( I.t -> I.t ) -> t -> t

val iter : ( I.t -> unit ) -> t -> t

val translate : int -> int -> t -> t

val viewport : ?x:int -> ?y:int -> int -> int -> t -> t

val bounding_box : t -> (int * int) * (int * int)

val scale : ?scaler:( I.t -> I.t -> unit ) -> (int * int) -> (int * int) -> t -> t

val resize : ?proportional:bool -> ?scaler:( I.t -> I.t -> unit ) -> int -> int -> t -> t

module Draw : sig ... end