Module Synth.Multitrack

Multichannel synthesizers.

class type t = object ... end

A multichannel synthesizer.

class create : int -> (int -> synth) -> t

Create a multichannel synthesizer with given number of channels and a function returning the synthesizer on each channel.