Module Audio.Effect

Audio effects.

class type t = object ... end

A possibly stateful audio effect.

class chain : t -> t -> t

class of_mono : int -> (unit -> Mono.Effect.t) -> t

class type delay_t = object ... end

val delay : int -> int -> float -> ?once:bool -> ?ping_pong:bool -> float -> delay_t

delay chans samplerate d once feedback creates a delay operator for buffer with chans channels at samplerate samplerate with d as delay in seconds and feedback as feedback. If once is set to true only one echo will be heard (no feedback).

class compress : ?attack:float -> ?release:float -> ?threshold:float -> ?ratio:float -> ?knee: float -> ?rms_window:float -> ?gain:float -> int -> int -> object ... end

Hardknee compressor with RMS look-ahead envelope calculation and adjustable attack/decay. Given parameters are attack and release in seconds, ratio n means n:1 compression, threshold and knee in dB, and rms_window in second is the duration for RMS acquisition. gain is an additional pre-gain.

class biquad_filter : int -> int -> [ `Band_pass | `High_pass | `Low_pass | `Notch | `All_pass | `Peaking | `Low_shelf | `High_shelf ] -> ?gain:float -> float -> float -> t

A biquadratic filter. gain in dB is only used by peaking, low and high shelves.

val auto_gain_control : int -> int -> ?rms_target:float -> ?rms_window:float -> ?kup:float -> ?kdown:float -> ?rms_threshold:float -> ?volume_init:float -> ?volume_min:float -> ?volume_max:float -> unit -> t