Module Audio.IO

Operation for reading and writing audio data from files, streams or devices.

exception Invalid_file

The file is not valid.

exception Invalid_operation

The operation is not valid on the file/device.

exception End_of_stream

Trying to read past the end of the stream.

module Reader : sig ... end

module Writer : sig ... end

module RW : sig ... end